Rename FrontpageFlarumStyle to FlarumStyle
