+++ Sources/LightPortal/addons/FlarumStyle/langs/english.php
@@ -1,0 +1,6 @@
+<?php
+
+$txt['lp_flarum_style_description'] = 'This addon will make a design of the frontpage articles like the style of the Flarum engine.<br>Works if forum topics are displayed on the portal frontpage.';
+
+$txt['lp_flarum_style_addon_all_boards'] = 'All boards';
+$txt['lp_flarum_style_addon_tags']       = 'Tags';
