+++ Sources/LightPortal/addons/FlarumStyle/langs/ukrainian.php
@@ -1,0 +1,6 @@
+<?php
+
+$txt['lp_flarum_style_description'] = 'Оформлення списку статей на головній сторінці порталу в стилі движка Flarum.<br>Працює, якщо на першій сторінці порталу відображаються теми форуму.';
+
+$txt['lp_flarum_style_addon_all_boards'] = 'Всі розділи';
+$txt['lp_flarum_style_addon_tags']       = 'Теги';
