--- Sources/LightPortal/addons/FrontpageFlarumStyle/langs/russian.php
@@ -1,6 +1,0 @@
-<?php
-
-$txt['lp_frontpage_flarum_style_description'] = 'Оформление списка статей на главной странице портала в стиле движка Flarum.<br>Работает, если на первой странице портала отображаются темы форума.';
-
-$txt['lp_frontpage_flarum_style_addon_all_boards'] = 'Все разделы';
-$txt['lp_frontpage_flarum_style_addon_tags']       = 'Теги';
