--- Sources/LightPortal/addons/FrontpageFlarumStyle/langs/ukrainian.php
@@ -1,6 +1,0 @@
-<?php
-
-$txt['lp_frontpage_flarum_style_description'] = 'Оформлення списку статей на головній сторінці порталу в стилі движка Flarum.<br>Працює, якщо на першій сторінці порталу відображаються теми форуму.';
-
-$txt['lp_frontpage_flarum_style_addon_all_boards'] = 'Всі розділи';
-$txt['lp_frontpage_flarum_style_addon_tags']       = 'Теги';
